feat(cmd): drop install/uninstall from CLI
* Also clean any deadcode mostly due to this change
