+++ pkg/util/watch/watch.go
@@ -23,7 +23,6 @@ import (
 
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/watch"
 
 	v1 "github.com/apache/camel-k/v2/pkg/apis/camel/v1"
 	"github.com/apache/camel-k/v2/pkg/client"
@@ -118,125 +117,6 @@ func HandleIntegrationEvents(ctx context.Context, c client.Client, integration *
 					}
 				}
 			}
-		}
-	}
-}
-
-// HandlePlatformStateChanges watches a platform resource and invoke the given handler when its status changes.
-// This function blocks until the handler function returns true or either the events channel or the context is closed.
-func HandlePlatformStateChanges(ctx context.Context, c client.Client, platform *v1.IntegrationPlatform, handler func(platform *v1.IntegrationPlatform) bool) error {
-	watcher, err := c.CamelV1().IntegrationPlatforms(platform.Namespace).
-		Watch(ctx, metav1.ListOptions{
-			FieldSelector: "metadata.name=" + platform.Name,
-		})
-	if err != nil {
-		return err
-	}
-
-	defer watcher.Stop()
-	events := watcher.ResultChan()
-
-	var lastObservedState *v1.IntegrationPlatformPhase
-
-	handlerWrapper := func(pl *v1.IntegrationPlatform) bool {
-		if lastObservedState == nil || *lastObservedState != pl.Status.Phase {
-			lastObservedState = &pl.Status.Phase
-			if !handler(pl) {
-				return false
-			}
-		}
-		return true
-	}
-
-	// Check completion before starting the watch
-	if !handlerWrapper(platform) {
-		return nil
-	}
-
-	for {
-		select {
-		case <-ctx.Done():
-			return nil
-		case e, ok := <-events:
-			if !ok {
-				return nil
-			}
-			if e.Object != nil {
-				if p, ok := e.Object.(*v1.IntegrationPlatform); ok {
-					if !handlerWrapper(p) {
-						return nil
-					}
-				}
-			}
-		}
-	}
-}
-
-// HandleIntegrationPlatformEvents watches all events related to the given integration platform.
-// This function blocks until the handler function returns true or either the events channel or the context is closed.
-//
-//nolint:nestif
-func HandleIntegrationPlatformEvents(ctx context.Context, c client.Client, p *v1.IntegrationPlatform,
-	handler func(event *corev1.Event) bool) error {
-	watcher, err := c.CoreV1().Events(p.Namespace).
-		Watch(ctx, metav1.ListOptions{
-			FieldSelector: fmt.Sprintf("involvedObject.kind=IntegrationPlatform,"+
-				"involvedObject.apiVersion=%s,"+
-				"involvedObject.name=%s",
-				v1.SchemeGroupVersion.String(), p.Name),
-		})
-	if err != nil {
-		return err
-	}
-
-	defer watcher.Stop()
-	events := watcher.ResultChan()
-
-	var lastEvent *corev1.Event
-	for {
-		select {
-		case <-ctx.Done():
-			return nil
-		case e, ok := <-events:
-			if !ok {
-				return nil
-			}
-
-			if e.Object != nil {
-				if evt, ok := e.Object.(*corev1.Event); ok {
-					if isAllowed(lastEvent, evt, p.CreationTimestamp.UnixNano()) {
-						lastEvent = evt
-						if !handler(evt) {
-							return nil
-						}
-					}
-				}
-			}
-		}
-	}
-}
-
-// WaitPodToTerminate will wait for a given pod to teminate.
-func WaitPodToTerminate(ctx context.Context, c client.Client, pod *corev1.Pod) error {
-	opts := metav1.ListOptions{
-		TypeMeta:      metav1.TypeMeta{},
-		FieldSelector: fmt.Sprintf("metadata.name=%s", pod.Name),
-	}
-	watcher, err := c.CoreV1().Pods(pod.Namespace).Watch(ctx, opts)
-	if err != nil {
-		return err
-	}
-
-	defer watcher.Stop()
-
-	for {
-		select {
-		case event := <-watcher.ResultChan():
-			if event.Type == watch.Deleted {
-				return nil
-			}
-		case <-ctx.Done():
-			return nil
 		}
 	}
 }
